+++ sdk/communication/Azure.Communication.CallAutomation/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Release History
 
 ## 1.1.0-beta.1 (Unreleased)
+- Play and recognize supports TTS and SSML source prompts.
+- Recognize supports choices and freeform speech.
+- Start/Stop continuous DTMF recognition by subscribing/unsubscribing to tones.
+- Send DTMF tones to a participant in the call.
+- Mute participants in the call.
 
 ### Features Added